Update create.html
updated categories to match up with index.html. removed the image preview (feel free to put it back in if you think it can genuinely be done, otherwise it's just a submit button now). feel free to pretty this page up, but keep the dropdown fields the same as they match up with db enums